requirements.txt
grandalf>=0.8
langchain>=1.2.3
langchain-community>=0.4.1
langchain-core>=1.2.7
langchain-google-genai>=4.1.3
langchain-groq>=1.1.1
langchain-huggingface>=1.2.0
langchain-openai>=1.1.7
python-dotenv>=1.2.1
scikit-learn>=1.8.0
typing>=3.10.0.0
